+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 See also the [closure compiler changelog](https://github.com/google/closure-compiler/wiki/Releases).
 
+## 2.8.0 (work-in-progress)
+
+- Update to closure compiler version `v20190709`
+- Fix #50
+
 ## 2.7.0
 
 - Update to closure compiler version `v20190618`